+++ src/FSharp.Formatting.Common/Range.fs
@@ -4,9 +4,10 @@ type MarkdownRange = { StartLine : int; StartColumn : int; EndLine : int; EndCol
 
 [<CompilationRepresentation (CompilationRepresentationFlags.ModuleSuffix)>]
 module MarkdownRange =
-    let Zero = { StartLine = 0; StartColumn = 0; EndLine = 0; EndColumn = 0 }
+    let zero = { StartLine = 0; StartColumn = 0; EndLine = 0; EndColumn = 0 }
 
-    let MergeRanges (ranges:MarkdownRange list) =
-        let startRange = ranges |> List.minBy (fun r -> r.StartLine * 10 + r.StartColumn)
-        let endRange = ranges |> List.maxBy (fun r -> r.EndLine * 10 + r.EndColumn)
-        { StartLine = startRange.StartLine; StartColumn = startRange.StartColumn; EndLine = endRange.EndLine; EndColumn = endRange.EndColumn }
+    let mergeRanges (ranges:MarkdownRange list) =
+        let startRange = ranges |> List.minBy (fun r -> r.StartLine, r.StartColumn)
+        let endRange = ranges |> List.maxBy (fun r -> r.EndLine, r.EndColumn)
+        { StartLine = startRange.StartLine; StartColumn = startRange.StartColumn; 
+          EndLine = endRange.EndLine; EndColumn = endRange.EndColumn }